cppcheck: Fix "int returned as long" warning.
    warning: int result is returned as long value. If the return value
    is long to avoid loss of information, then you have loss of
    information. [truncLongCastReturn]

This function returns "unsigned long long" which cppcheck appears to
be misinterpreting.  Change the return type to uint64_t to eliminate
the warning.

+++ mythtv/libs/libmythtv/mpeg/dvbdescriptors.h
@@ -837,7 +837,7 @@ class SatelliteDeliverySystemDescriptor : public MPEGDescriptor
         return ((m_data[2]<<24) | (m_data[3]<<16) |
                 (m_data[4]<<8)  | (m_data[5]));
     }
-    unsigned long long FrequencykHz(void) const
+    uint64_t FrequencykHz(void) const
     {
         return byte4BCD2int(m_data[2], m_data[3], m_data[4], m_data[5]) * 10;
     }
